requirements.txt
anyio==3.5.0
asgiref==3.5.0
certifi==2021.10.8
cffi==1.15.0
click==8.0.4
cryptography==36.0.1
fastapi==0.74.1
greenlet==1.1.2
h11==0.13.0
httptools==0.3.0
idna==3.3
Jinja2==3.0.3
jwt==1.3.1
MarkupSafe==2.1.0
pycparser==2.21
pydantic==1.9.0
PyMySQL==1.0.2
python-dotenv==0.19.2
python-multipart==0.0.5
PyYAML==6.0
six==1.16.0
sniffio==1.2.0
SQLAlchemy==1.4.31
starlette==0.17.1
typing_extensions==4.1.1
uvicorn==0.17.5
uvloop==0.16.0
watchgod==0.7
websockets==10.2
